fix gson bug and write code for ui components

+++ app/src/main/java/com/mehranj73/moviedb/data/model/Movie.kt
@@ -18,6 +18,6 @@ data class Movie(
     val poster_path: String,
     val release_date: String,
     val title: String,
-    val vote_average: Int,
+    val vote_average: Float,
     val vote_count: Int
 )
